Revert "Merge branch 'feature/insurance-claim' into feature/insurance-claim-details"
This reverts commit 72de9e598016e43b6849fe540f68b04cb3197921, reversing
changes made to 3096abc7329d0be95790f5c1aad352128bd96d9e.

+++ src/main/java/com/azid/auth/backend/AZ/Auth/model/Plan.java
@@ -5,8 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -26,12 +24,12 @@ public class Plan {
     private Double coverageAmount;
 
     @Column(name = "base_premium")
-    private BigDecimal basePremium;
+    private Double basePremium;
 
     @Column(name = "duration")
     private Integer duration;
 
     @Column(name = "status")
-    private String status; // ACTIVE / INACTIVE
+    private String status;
 
 }
